fix(desktop): place new v2 workspace at top of sidebar
The revert of #4120 (#4135) restored a hardcoded `tabOrder: 0` in the
v2 workspace.create path. Existing rows default to `tabOrder: 0` too,
so new workspaces tied with them and ordering was non-deterministic.

Restores the `getPrependTabOrder` helper from #4120 as a shared util in
`dashboardSidebarLocal/`, and uses it from `useWorkspaceCreates` so new
rows land strictly above every existing top-level item — matching what
the pending-row injection in `useDashboardSidebarData` already assumes.
