+++ apps/desktop/src/renderer/routes/_authenticated/providers/CollectionsProvider/dashboardSidebarLocal/index.ts
@@ -1,2 +1,3 @@
 export * from "./schema";
 export * from "./sidebarVisibility";
+export * from "./tabOrder";
