+++ apps/desktop/src/renderer/stores/workspace-creates/useWorkspaceCreates.ts
@@ -5,6 +5,10 @@ import { authClient } from "renderer/lib/auth-client";
 import { getHostServiceClientByUrl } from "renderer/lib/host-service-client";
 import type { PaneViewerData } from "renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/types";
 import { useCollections } from "renderer/routes/_authenticated/providers/CollectionsProvider";
+import {
+	getPrependTabOrder,
+	isSidebarWorkspaceVisible,
+} from "renderer/routes/_authenticated/providers/CollectionsProvider/dashboardSidebarLocal";
 import { useLocalHostService } from "renderer/routes/_authenticated/providers/LocalHostServiceProvider";
 import { appendLaunchesToPaneLayout } from "./appendLaunchesToPaneLayout";
 import {
@@ -82,12 +86,26 @@ export function useWorkspaceCreates(): UseWorkspaceCreatesApi {
 						},
 					);
 				} else {
+					const projectId = result.workspace.projectId;
+					const topLevelItems = [
+						...Array.from(collections.v2WorkspaceLocalState.state.values())
+							.filter(
+								(item) =>
+									item.sidebarState.projectId === projectId &&
+									item.sidebarState.sectionId === null &&
+									isSidebarWorkspaceVisible(item),
+							)
+							.map((item) => ({ tabOrder: item.sidebarState.tabOrder })),
+						...Array.from(collections.v2SidebarSections.state.values())
+							.filter((item) => item.projectId === projectId)
+							.map((item) => ({ tabOrder: item.tabOrder })),
+					];
 					collections.v2WorkspaceLocalState.insert({
 						workspaceId: result.workspace.id,
 						createdAt: new Date(),
 						sidebarState: {
-							projectId: result.workspace.projectId,
-							tabOrder: 0,
+							projectId,
+							tabOrder: getPrependTabOrder(topLevelItems),
 							sectionId: null,
 							changesFilter: { kind: "all" },
 							activeTab: "changes",
